Gradient Generator
This project is a Gradient Generator built using HTML, CSS, and JavaScript that allows users to create and preview beautiful linear gradients in real time. Users can choose their own colors, see the live result, and copy the generated CSS code for use in web design and development. It's a great utility for frontend developers, designers, and anyone looking to add creative background styles to their websites. This project is part of my #100DaysOfCode challenge to sharpen my skills in real-world web development.
#HTML #CSS #JavaScript #WebDevelopment #FrontendDevelopment #GradientGenerator #UIUX #OpenSource #CodingChallenge #DeveloperTools #MadeWithCode #100DaysOfCode #AnamikaAnand